+++ Sources/Scru128/Scru128Id.swift
@@ -200,14 +200,33 @@ extension Scru128Id: Codable {
     try container.encode(description)
   }
 
-  /// Decodes the object from a 25-digit canonical string representation.
+  /// Decodes the object from a 25-digit canonical string representation or  a 16-byte big-endian
+  /// byte array.
   public init(from decoder: Decoder) throws {
     let container = try decoder.singleValueContainer()
-    if let bs = Self.parse(try container.decode(String.self)) {
-      bytes = bs
+    if let strValue = try? container.decode(String.self) {
+      guard let bs = Self.parse(strValue) else {
+        throw DecodingError.dataCorruptedError(
+          in: container, debugDescription: "could not parse string as Scru128Id")
+      }
+      self.init(bs)
+    } else if let byteArray = try? container.decode([UInt8].self) {
+      if byteArray.count == 16 {
+        self.init(byteArray)
+      } else {
+        guard let strValue = String(bytes: byteArray, encoding: .utf8) else {
+          throw DecodingError.dataCorruptedError(
+            in: container, debugDescription: "could not parse byte array as Scru128Id")
+        }
+        guard let bs = Self.parse(strValue) else {
+          throw DecodingError.dataCorruptedError(
+            in: container, debugDescription: "could not parse byte array as Scru128Id")
+        }
+        self.init(bs)
+      }
     } else {
       throw DecodingError.dataCorruptedError(
-        in: container, debugDescription: "could not parse string as Scru128Id")
+        in: container, debugDescription: "expected string or byte array but found neither")
     }
   }
 }